refactor: rename the file format to snake_case
