+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/DialogBackgroundView.tsx
@@ -12,7 +12,7 @@ import {
 	MenuImages,
 	MenuImagesFooter,
 	MenuSearch,
-} from "./style";
+} from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/dialog-background.style";
 import { IDialogBackgroundView } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/types";
 
 const DialogBackgroundView: React.FC<IDialogBackgroundView> = props => {
