+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/components/MessageError/types/message-error.types.ts
@@ -1,6 +1,6 @@
 import { IImages } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/types";
-import { IFetchSearchImagesResponse } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types";
+import { IGetSearchImagesResponse } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types";
 
 export interface IMessageError {
-	images: IImages | IFetchSearchImagesResponse;
+	images: IImages | IGetSearchImagesResponse;
 }
