+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/components/MessageNoResultsFound/types/message-no-results-found.types.ts
@@ -1,6 +1,6 @@
 import { IImages } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/types";
-import { IFetchSearchImagesResponse } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types";
+import { IGetSearchImagesResponse } from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types";
 
 export interface IMessageNoResultsFound {
-	images: IImages | IFetchSearchImagesResponse;
+	images: IImages | IGetSearchImagesResponse;
 }
