--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./DialogBackground.constants";
