+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/dialog-background.constants.ts
