+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/get-search-image.service.ts
@@ -1,0 +1,13 @@
+import {
+	IGetSearchImageService,
+	IGetSearchImagesResponse,
+} from "modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types";
+import api from "shared/services/api";
+
+export const getSearchImageService: IGetSearchImageService = async ({
+	params,
+}) => {
+	return await api.get<IGetSearchImagesResponse>("images/search", {
+		params,
+	});
+};
