+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/index.ts
@@ -1,1 +1,1 @@
-export * from "./fetchGetSearchImageService";
+export * from "./get-search-image.service";
