+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./fetchGetSearchImageService.types";
+export * from "./get-search-image.types";
