--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/style/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./DialogBackground.style";
-export * from "./DialogBackground.style";
