--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/style/types/DialogBackground.style.types.ts
@@ -1,6 +1,0 @@
-import { Theme } from "@emotion/react";
-
-export interface IDialogBackgroundStyled {
-	theme?: Theme;
-	backgroundimage: string;
-}
