--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/style/types/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./DialogBackground.style.types";
-export * from "./DialogBackground.style.types";
