+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/types/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./DialogBackground.types";
-export * from "./DialogBackground.types";
+export * from "./dialog-background.types";
