--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/constants/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./DialogEditBoard.constants";
-export * from "./DialogEditBoard.constants";
