--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/schema/DialogEditBoard.schema.ts
@@ -1,8 +1,0 @@
-import * as Yup from "yup";
-
-export const DialogEditBoardSchema = Yup.object({
-	title: Yup.string()
-		.required()
-		.min(3, "Should be 3 chars minimum")
-		.max(255, "Max of 255 characters"),
-});
