--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/schema/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./DialogEditBoard.schema";
-export * from "./DialogEditBoard.schema";
