+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/services/index.ts
@@ -1,1 +1,1 @@
-export * from "./fetchPatchBoardService";
+export * from "./patch-board.service";
