+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/services/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./fetchPatchBoardService.types";
+export * from "./patch-board.types";
