--- src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/style/types/DialogBackground.style.ts
@@ -1,6 +1,0 @@
-import { Theme } from "@emotion/react";
-
-export interface IDialogBackgroundStyled {
-	theme?: Theme;
-	backgroundimage: string;
-}
