+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/types/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./DialogEditBoard.types";
-export * from "./DialogEditBoard.types";
+export * from "./dialog-edit-board.types";
